Add role templating support
The default template used by ansible-galaxy is good, but it doesn't
allow organizations to customize it to suit their needs.  This patch
addresses that by adding the option to point ansible-galaxy at a
template directory.  The contents of this directory are then
copied (or rendered) into the output role directory.  Several command
line options and configuration entries are added to allow for further
customization from the template.