Move cli options to vars file
Use a yaml file to set variables like company, license, etc rather than
using command line options.

Author is pulled out as a separate configuration parameter to allow the
vars file for templates to be more easily shared between groups or
developers.  This also allows for the author to be set in an environment
variable.

The configuration parameter called role_template_ignore can be used to
prune files or directories from the template.  The defaults ignore .git
and .git_keep to allow for the role template to be a git repository.
The variable will accept a comma separated list of regular expressions.
However, the regex will be tested against all files and directories
without any regard to whether it is a file or directory.  The current
implementation is also unaware of the path to the file/directory being
tested.  Some improvements may be in order here.