Support for user defined variables at role init
This extends the new role skeleton functionality by allowing for a
init_vars_file to update the default set of variables that are injected
into the templates from the skeleton.  In addition to the vars file, an
author can be injected via config or env_var, which should make the
init_vars_file more portable.

A couple use cases:
* An orginization with multiple sub units can share a common role
  skeleton.  Using templates to customize it for each subunit.
* Those satisfied with the default skeleton could use this feature to
  populate author/company/etc metadata without the need to build a
  custom role skeleton.